Finished CRUD operations, fleshed out note editing UI

+++ app/src/main/java/com/example/notes/infrastructure/persistence/daos/NoteDAO.kt
@@ -15,7 +15,7 @@ interface NoteDAO {
     fun getNotes(): Flow<List<Note>>
 
     @Query("SELECT * FROM note WHERE id = :id")
-    suspend fun getNoteById(id: Int): Note?
+    suspend fun getNoteById(id: Long): Note?
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertNote(note: Note)
